apps/backend/requirements.txt
fastapi
uvicorn 
sqlalchemy
asyncpg 
pytest 
httpx
python-dotenv 